Move the tooltip into better namespace.
